+++ src/livecodes/editor/monaco/monaco.ts
@@ -234,8 +234,9 @@ export const createEditor = async (options: EditorOptions): Promise<CodeEditor> 
     monaco: typeof Monaco;
   }): CompilerOptions => {
     const settings: CompilerOptions = {
-      checkJs: !isJSLang,
       allowJs: isJSLang,
+      checkJs: !isJSLang,
+      strictNullChecks: !isJSLang,
       allowNonTsExtensions: true,
       experimentalDecorators: true,
       emitDecoratorMetadata: true,
@@ -284,7 +285,7 @@ export const createEditor = async (options: EditorOptions): Promise<CodeEditor> 
     ];
     const isJSLang = JSLangs.includes(language);
     const isJsx = hasJsx.includes(language);
-    const nonReactJsx = language === 'stencil';
+    const nonReactJsx = ['solid', 'solid.tsx', 'stencil'].includes(language);
 
     if (
       !['script', 'tests', 'editorSettings'].includes(editorId) ||